+++ WebAutomation/src/test/java/Runner/RunnerTest.java
@@ -18,9 +18,9 @@ import net.masterthought.cucumber.ReportBuilder;
 import net.masterthought.cucumber.sorting.SortingMethod;
 
 	@RunWith(Cucumber.class)
-	@CucumberOptions(features= "src/test/resources/Features/frontend/SmokeTest.feature",
+	@CucumberOptions(features= "src/test/resources/Features/frontend",
 	glue={"steps"},	
-	tags="@test",
+	//tags="@test",
 	plugin = {"com.aventstack.extentreports.cucumber.adapter.ExtentCucumberAdapter:",
 			  "pretty","json:target/cucumber-reports/Cucumber.json", 
 			  "junit:target/cucumber-reports/Cucumber.xml",
