+++ WebAutomation/src/test/java/steps/frontend/studiodetailpagesteps.java
@@ -1,0 +1,36 @@
+package steps.frontend;
+
+import org.apache.log4j.Logger;
+import org.openqa.selenium.support.ui.WebDriverWait;
+
+import Pageobjects.frontend.ShareFeature;
+import Pageobjects.frontend.ToastandErrormessages;
+import Pageobjects.frontend.commonlocatorsandmethods;
+import Pageobjects.frontend.homepage;
+import Pageobjects.frontend.paymentpage;
+import Pageobjects.frontend.shodetailpage;
+import Pageobjects.frontend.studiodetailpage;
+import Pageobjects.frontend.videoplayer;
+import Resources.BaseSetup;
+import io.cucumber.java.en.Then;
+
+public class studiodetailpagesteps  extends BaseSetup{
+	
+	homepage home=new homepage();
+	shodetailpage detailpage=new shodetailpage();
+	paymentpage payment=new paymentpage();
+	ToastandErrormessages toaster=new ToastandErrormessages();
+	commonlocatorsandmethods cm=new commonlocatorsandmethods();
+	ShareFeature share=new ShareFeature();
+	studiodetailpage studio=new studiodetailpage();
+	public static Logger log=Logger.getLogger(studiodetailpagesteps.class.getName());
+	videoplayer video=new videoplayer();
+	WebDriverWait wait=new WebDriverWait(driver,30);
+	
+	
+	 @Then("^Share studio with all share icons$")
+	    public void share_studio_with_all_share_icons() throws Throwable {
+	        ShareFeature.Studioshare();
+	    }
+
+}
